pyproject.toml
[project]
name = "linkedin-post-bot"
version = "0.1.0"
description = "Generate a LinkedIn post of about 250 words for a given topic and context"
requires-python = ">=3.9"
dependencies = [
    "pyperclip",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
